Make arguments of the `add_*_argument` functions keyword-only
Only the `parser`, `short_alias`, and `long_alias` arguments remain
positional. This is in sync with how `ArgumentParser.add_argument`
is normally used (where only the short/long option names are
positional).

The only function that is expected to be really affected by the
change is `add_version_argument`, where the `version` argument
could be (and was often) used as positional.